Add files via upload
In canteen, people order food, make payment and get their order. These all things are
done manually. Canteen workers use coupons for each medium order or big order. So that at
the end of the day they can keep record of all the orders and payments in register which is
maintained by canteen manager. But the problem is for every small order, coupon system is
not there. These order records are done manually in mind. So there may be chances of data
loss. For example, if there are total 30 orders for coffee in a day but if by some manual
mistake only 29 orders are recorded, so wrong entries are made for coffee. This may happen
for more than one food items. The canteen management system has been developed to
override the prevailing practicing in manual system. This software is supported to eliminate
& in some cases reduce hardships faced by existing system. Moreover, this system is
designed for the particular need of the canteen to carry out operations in smooth and effective manner.